refactor: remove functions
