+++ packages/sheets-ui/src/commands/commands/add-worksheet-merge.command.ts
@@ -1,4 +1,4 @@
-import type { ICellData, ICommand, IMutationInfo, IRange, Worksheet } from '@univerjs/core';
+import type { ICommand, IMutationInfo, IRange } from '@univerjs/core';
 import {
     CommandType,
     Dimension,
@@ -6,14 +6,9 @@ import {
     IUndoRedoService,
     IUniverInstanceService,
     LocaleService,
-    ObjectMatrix,
     sequenceExecute,
 } from '@univerjs/core';
-import type {
-    IAddWorksheetMergeMutationParams,
-    IRemoveWorksheetMergeMutationParams,
-    ISetRangeValuesMutationParams,
-} from '@univerjs/sheets';
+import type { IAddWorksheetMergeMutationParams, IRemoveWorksheetMergeMutationParams } from '@univerjs/sheets';
 import {
     AddMergeUndoMutationFactory,
     AddWorksheetMergeMutation,
@@ -21,87 +16,17 @@ import {
     RemoveMergeUndoMutationFactory,
     RemoveWorksheetMergeMutation,
     SelectionManagerService,
-    SetRangeValuesMutation,
-    SetRangeValuesUndoMutationFactory,
 } from '@univerjs/sheets';
 import { IConfirmService } from '@univerjs/ui';
 import type { IAccessor } from '@wendellhu/redi';
+
+import { checkCellContentInRanges, getClearContentMutationParamsForRanges } from '../../common/utils';
 
 export interface IAddMergeCommandParams {
     value?: Dimension.ROWS | Dimension.COLUMNS;
     selections: IRange[];
     workbookId: string;
     worksheetId: string;
-}
-
-export function checkCellContentInRanges(worksheet: Worksheet, ranges: IRange[]): boolean {
-    return ranges.some((range) => checkCellContentInRange(worksheet, range));
-}
-
-export function checkCellContentInRange(worksheet: Worksheet, range: IRange): boolean {
-    const { startRow, startColumn, endColumn, endRow } = range;
-    const cellMatrix = worksheet.getMatrixWithMergedCells(startRow, startColumn, endRow, endColumn);
-
-    let someCellGoingToBeRemoved = false;
-    cellMatrix.forValue((row, col, cellData) => {
-        if (cellData && (row !== startRow || col !== startColumn) && worksheet.cellHasValue(cellData)) {
-            someCellGoingToBeRemoved = true;
-            return false;
-        }
-    });
-    return someCellGoingToBeRemoved;
-}
-
-export function getClearContentMutationParamsForRanges(
-    accessor: IAccessor,
-    workbookId: string,
-    worksheet: Worksheet,
-    ranges: IRange[]
-): {
-    undos: IMutationInfo[];
-    redos: IMutationInfo[];
-} {
-    const undos: IMutationInfo[] = [];
-    const redos: IMutationInfo[] = [];
-
-    const worksheetId = worksheet.getSheetId();
-
-    // Use the following file as a reference.
-    // packages/sheets/src/commands/commands/clear-selection-all.command.ts
-    // packages/sheets/src/commands/mutations/set-range-values.mutation.ts
-    ranges.forEach((range) => {
-        const redoMatrix = getClearContentMutationParamForRange(worksheet, range);
-        const redoMutationParams: ISetRangeValuesMutationParams = {
-            workbookId,
-            worksheetId,
-            cellValue: redoMatrix.getData(),
-        };
-        const undoMutationParams: ISetRangeValuesMutationParams = SetRangeValuesUndoMutationFactory(
-            accessor,
-            redoMutationParams
-        );
-
-        undos.push({ id: SetRangeValuesMutation.id, params: undoMutationParams });
-        redos.push({ id: SetRangeValuesMutation.id, params: redoMutationParams });
-    });
-
-    return {
-        undos,
-        redos,
-    };
-}
-
-export function getClearContentMutationParamForRange(worksheet: Worksheet, range: IRange): ObjectMatrix<ICellData> {
-    const { startRow, startColumn, endColumn, endRow } = range;
-    const cellMatrix = worksheet.getMatrixWithMergedCells(startRow, startColumn, endRow, endColumn);
-    const redoMatrix = new ObjectMatrix<ICellData>();
-    cellMatrix.forValue((row, col, cellData) => {
-        if (cellData && (row !== startRow || col !== startColumn)) {
-            redoMatrix.setValue(row, col, null);
-        }
-    });
-
-    return redoMatrix;
 }
 
 export const AddWorksheetMergeCommand: ICommand = {
